rename folder and add refetchs on every mutations that has an impact on the draft
