+++ platform/web/packages/web-app/src/App/pages/data/Draft/DraftPage/ValidatedDraftInfo.tsx
