+++ platform/web/packages/web-app/src/App/pages/data/Draft/DraftPage/useDraftMutations.ts
@@ -120,6 +120,7 @@ export const useDraftMutations = (params: useDraftMutationsParams) => {
       if (res) {
         queryClient.invalidateQueries({ queryKey: ["data/catalogueGet", { id: catalogueId! }] })
         queryClient.invalidateQueries({ queryKey: ["data/catalogueGetByIdentifier", { identifier: catalogueId! }] })
+        queryClient.invalidateQueries({ queryKey: ["data/catalogueDraftGet", { id: draftId }] })
         queryClient.invalidateQueries({ queryKey: ["data/catalogueDraftPage"] })
         navigate(cataloguesContributions())
       }
@@ -137,6 +138,7 @@ export const useDraftMutations = (params: useDraftMutationsParams) => {
         if (res) {
           queryClient.invalidateQueries({ queryKey: ["data/catalogueGet", { id: catalogueId! }] })
           queryClient.invalidateQueries({ queryKey: ["data/catalogueGetByIdentifier", { identifier: catalogueId! }] })
+          queryClient.invalidateQueries({ queryKey: ["data/catalogueDraftGet", { id: draftId }] })
           queryClient.invalidateQueries({ queryKey: ["data/catalogueDraftPage"] })
           queryClient.invalidateQueries({ queryKey: ["data/cataloguePage"] })
           queryClient.invalidateQueries({ queryKey: ["data/catalogueRefGetTree"] })
