+++ platform/web/packages/web-app/src/App/pages/data/Draft/DraftPage/useDraftValidations.ts
@@ -74,6 +74,7 @@ export const useDraftValidations = (params: useDraftValidationsParams) => {
         refetchDraft()
         queryClient.invalidateQueries({ queryKey: ["data/catalogueGet", { id: draft.catalogue.id }] })
         queryClient.invalidateQueries({ queryKey: ["data/catalogueGetByIdentifier", { id: draft.catalogue.identifier }] })
+        queryClient.invalidateQueries({ queryKey: ["data/catalogueDraftGet", { id: draftId }] })
         queryClient.invalidateQueries({ queryKey: ["data/catalogueDraftPage"] })
         navigate(cataloguesContributions() + "?successfullContribution=true")
       }
@@ -92,6 +93,7 @@ export const useDraftValidations = (params: useDraftValidationsParams) => {
       })
       if (res) {
         queryClient.invalidateQueries({ queryKey: ["data/catalogueDraftPage"] })
+        queryClient.invalidateQueries({ queryKey: ["data/catalogueDraftGet", { id: draftId }] })
         navigate(cataloguesToVerify())
       }
     },
